+++ establishment-ms/src/main/java/com/codexmind/establishment/dto/MenuDTO.java
@@ -11,6 +11,6 @@ public class MenuDTO {
 
     private String name;
 
-    public Long establishment;
+    public Integer establishment;
 
 }
