+++ establishment-ms/src/main/java/com/codexmind/establishment/usecases/menu/DetailMenu.java
@@ -15,7 +15,7 @@ public class DetailMenu {
 
     private MenuRepository repository;
 
-    public Menu execute(Long id){
+    public Menu execute(Integer id){
         var menu = repository.findByIdAndStatus(id, Status.ACTIVE);
         return menu.orElseThrow(()-> new EntityNotFoundException(
                 "Object Not Found! id:  "+ id + " type: "+ Menu.class.getName()
